修复源码测试编译问题并添加样本验证程序
修复测试编译问题:
- 修正类型转换错误: int* -> size_t* (直方图和分水岭测试)
- 修正枚举名称: NPP_NORM_L2 -> nppiNormL2 (梯度测试)
- 修正分水岭API参数类型: Npp32s* -> Npp32u*

样本程序改进:
- 更新build.sh: 先构建OpenNPP库，再构建样本
- 修正CMakeLists.txt: 使用正确的库名称libopennpp.a
- 所有5个NPP样本现在都能成功编译和运行

添加验证程序:
- verify_results.cpp: 验证盒式滤波结果的正确性
- 实现数值验证和边缘处理验证
- 验证结果: ✅ 盒式滤波验证通过！

+++ test/unit/nppi/nppi_statistics_functions/test_nppi_histogram.cpp
@@ -56,7 +56,7 @@ TEST_F(NPPIHistogramTest, EvenLevelsHost_ErrorHandling) {
 // 测试直方图缓冲区大小获取
 TEST_F(NPPIHistogramTest, HistogramEvenGetBufferSize_8u_C1R) {
     int nLevels = 256;
-    int bufferSize = 0;
+    size_t bufferSize = 0;
     
     NppStatus status = nppiHistogramEvenGetBufferSize_8u_C1R(roi, nLevels, &bufferSize);
     EXPECT_EQ(status, NPP_SUCCESS);
@@ -80,7 +80,7 @@ TEST_F(NPPIHistogramTest, HistogramEven_8u_C1R_Basic) {
     std::vector<Npp32s> pHist(nLevels - 1);
     
     // 获取缓冲区大小
-    int bufferSize = 0;
+    size_t bufferSize = 0;
     NppStatus status = nppiHistogramEvenGetBufferSize_8u_C1R(roi, nLevels, &bufferSize);
     EXPECT_EQ(status, NPP_SUCCESS);
